refactor(wizards/connectedap): edit wizard and its imports
